+++ htdocs/components/RenewalsDashboard/subComponents/RenHostingComponent.js
@@ -1,7 +1,7 @@
 import MainComponent from "../../shared/MainComponent";
 import React from 'react';
 import Table from "../../shared/table/table";
-import {dateFormatExcludeNull, exportCSV} from "../../utils/utils";
+import {dateFormatExcludeNull, exportCSV, poundFormat} from "../../utils/utils";
 
 export class RenHostingComponent extends MainComponent {
     constructor(props) {
@@ -58,6 +58,26 @@ export class RenHostingComponent extends MainComponent {
                 sortable: true,
                 className: "text-center",
                 content: order => dateFormatExcludeNull(order.invoiceToDate),
+            },
+            {
+                path: "costAnnum",
+                label: "",
+                hdToolTip: "Cost Price/Annum",
+                hdClassName: "text-center",
+                icon: "fal fa-2x fa-coin color-gray2 pointer",
+                sortable: true,
+                className: "text-right",
+                content: order => poundFormat(order.costAnnum)
+            },
+            {
+                path: "saleAnnum",
+                label: "",
+                hdToolTip: "Sale Price/Annum",
+                hdClassName: "text-center",
+                icon: "fal fa-2x fa-coins color-gray2 pointer",
+                sortable: true,
+                className: "text-right",
+                content: order => poundFormat(order.saleAnnum)
             },
             {
                 path: "customerItemID",