+++ htdocs/components/RenewalsDashboard/subComponents/RenewalComponent.js
@@ -1,7 +1,7 @@
 import MainComponent from "../../shared/MainComponent";
 import React from 'react';
 import Table from "../../shared/table/table";
-import {dateFormatExcludeNull, exportCSV, sort} from "../../utils/utils";
+import {dateFormatExcludeNull, exportCSV, poundFormat, sort} from "../../utils/utils";
 
 export class RenewalComponent extends MainComponent {
     constructor(props) {
@@ -36,7 +36,6 @@ export class RenewalComponent extends MainComponent {
                 hdClassName: "text-center",
                 icon: "fal fa-2x fa-shopping-cart color-gray2 pointer",
                 sortable: true,
-                // className: "text-center",
                 backgroundColorColumn: "sentQuotationColor",
             },
             {
@@ -99,6 +98,28 @@ export class RenewalComponent extends MainComponent {
                 className: "text-center",
                 backgroundColorColumn: "sentQuotationColor",
                 content: order => dateFormatExcludeNull(order.latestQuoteSent, 'YYYY-MM-DD HH:mm:ss', 'DD/MM/YYYY HH:mm:ss')
+            },
+            {
+                path: "costAnnum",
+                label: "",
+                hdToolTip: "Cost Price/Annum",
+                hdClassName: "text-center",
+                icon: "fal fa-2x fa-coin color-gray2 pointer",
+                sortable: true,
+                className: "text-right",
+                backgroundColorColumn: "sentQuotationColor",
+                content: order => poundFormat(order.costAnnum)
+            },
+            {
+                path: "saleAnnum",
+                label: "",
+                hdToolTip: "Sale Price/Annum",
+                hdClassName: "text-center",
+                icon: "fal fa-2x fa-coins color-gray2 pointer",
+                sortable: true,
+                className: "text-right",
+                backgroundColorColumn: "sentQuotationColor",
+                content: order => poundFormat(order.saleAnnum)
             },
             {
                 path: "customerItemID",